+++ src/Symfony/Component/Translation/Extractor/ChainExtractor.php
@@ -48,6 +48,16 @@ class ChainExtractor implements ExtractorInterface
     /**
      * {@inheritdoc}
      */
+    public function blank(bool $blank): void
+    {
+        foreach ($this->extractors as $extractor) {
+            $extractor->blank($blank);
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function extract(string|iterable $directory, MessageCatalogue $catalogue)
     {
         foreach ($this->extractors as $extractor) {